requirements.txt
fastapi
uvicorn
google-generativeai
requests
elevenlabs
fpdf
python-dotenv
python-multipart